Beginnings of a tutorial
Add a tutorial/ directory and the beginnings of a tutorial.

Move the existing dummy.yaml file into tutorial (it forms part of
the first steps of the tutorial now).

Remove the shell_as_user test from dummy.yaml, which is not general
enough for many people to use (wildfly image doesn't contain a
dbecvari user)