Adjust module scanner exclusions
Exclude any path containing /tests or /. to avoid checking for modules
in tests (which can be noisy if e.g. mock is not installed in the run
environment) or .git etc.

Debug log if we skip something since it's a wider check and could be
hard to figure out why a directory was not being looked at.

Fixes #60